move tests to best-practices dir
