+++ best-practice/src/test/java/com/saucedemo/DesktopTests.java
@@ -1,4 +1,4 @@
-package com.asb;
+package com.saucedemo;
 
 import com.saucelabs.saucebindings.SauceOptions;
 import com.saucelabs.saucebindings.SauceSession;
